+++ weather@mockturtl/src/3_8/main.ts
@@ -6,7 +6,7 @@
 //----------------------------------------------------------------------
 
 import { DateTime } from "luxon";
-import { Config } from "./config";
+import { Config, ServiceClassMapping } from "./config";
 import { WeatherLoop } from "./loop";
 import { MetUk } from "./providers/met_uk";
 import { WeatherData, WeatherProvider, LocationData, AppletError, CustomIcons, NiceErrorDetail, RefreshState, BuiltinIcons } from "./types";
@@ -201,9 +201,9 @@ export class WeatherApplet extends TextIconApplet {
 
 	/** Displays weather info in applet's panel */
 	private DisplayWeather(weather: WeatherData): boolean {
-		let location = GenerateLocationText(weather, this.config);
+		const location = GenerateLocationText(weather, this.config);
 
-		let lastUpdatedTime = AwareDateString(weather.date, this.config.currentLocale, this.config._show24Hours, DateTime.local().zoneName);
+		const lastUpdatedTime = AwareDateString(weather.date, this.config.currentLocale, this.config._show24Hours, DateTime.local().zoneName);
 		this.SetAppletTooltip(`${location} - ${_("As of {lastUpdatedTime}", { "lastUpdatedTime": lastUpdatedTime })}`);
 		this.DisplayWeatherOnLabel(weather.temperature, weather.condition.description);
 		this.SetAppletIcon(weather.condition.icons, weather.condition.customIcon);
@@ -258,7 +258,7 @@ export class WeatherApplet extends TextIconApplet {
 			this.SetCustomIcon(customIcon);
 		}
 		else {
-			let icon = WeatherIconSafely(iconNames, this.config.AppletIconType);
+			const icon = WeatherIconSafely(iconNames, this.config.AppletIconType);
 			this.config.AppletIconType == IconType.SYMBOLIC ?
 				this.set_applet_icon_symbolic_name(icon) :
 				this.set_applet_icon_name(icon);
@@ -297,7 +297,7 @@ export class WeatherApplet extends TextIconApplet {
 	 * @param method default is GET
 	 */
 	public async LoadJsonAsync<T>(this: WeatherApplet, url: string, params?: HTTPParams, HandleError?: (message: HttpError) => boolean, headers?: HTTPHeaders, method: Method = "GET"): Promise<T | null> {
-		let response = await HttpLib.Instance.LoadJsonAsync<T>(url, params, headers, method);
+		const response = await HttpLib.Instance.LoadJsonAsync<T>(url, params, headers, method);
 
 		// We have errorData inside
 		if (!response.Success) {
@@ -321,7 +321,7 @@ export class WeatherApplet extends TextIconApplet {
 	 * @param method default is GET
 	 */
 	public async LoadAsync(this: WeatherApplet, url: string, params?: HTTPParams, HandleError?: (message: HttpError) => boolean, headers?: HTTPHeaders, method: Method = "GET"): Promise<string | null> {
-		let response = await HttpLib.Instance.LoadAsync(url, params, headers, method);
+		const response = await HttpLib.Instance.LoadAsync(url, params, headers, method);
 
 		// We have errorData inside
 		if (!response.Success) {
@@ -341,12 +341,12 @@ export class WeatherApplet extends TextIconApplet {
 	// Config Callbacks, do not delete
 
 	private async locationLookup(): Promise<void> {
-		let command = "xdg-open ";
+		const command = "xdg-open ";
 		spawnCommandLine(command + "https://cinnamon-spices.linuxmint.com/applets/view/17");
 	}
 
 	private async submitIssue(): Promise<void> {
-		let command = "xdg-open";
+		const command = "xdg-open";
 		const baseUrl = 'https://github.com/linuxmint/cinnamon-spices-applets/issues/new';
 		const title = "weather@mockturl - ";
 
@@ -469,9 +469,8 @@ The contents of the file saved from the applet help page goes here
 
 	/** Into right-click context menu */
 	private AddRefreshButton(): void {
-		let itemLabel = _("Refresh")
-		// () => functions do not need to bind context
-		let refreshMenuItem = new MenuItem(itemLabel, REFRESH_ICON, () => this.RefreshAndRebuild());
+		const itemLabel = _("Refresh")
+		const refreshMenuItem = new MenuItem(itemLabel, REFRESH_ICON, () => this.RefreshAndRebuild());
 		this._applet_context_menu.addMenuItem(refreshMenuItem);
 	}
 
@@ -480,7 +479,7 @@ The contents of the file saved from the applet help page goes here
 	 * @param error 
 	 */
 	private HandleHTTPError(error: HttpError): void {
-		let appletError: AppletError = {
+		const appletError: AppletError = {
 			detail: error.message,
 			userError: false,
 			code: error.code,
@@ -506,39 +505,9 @@ The contents of the file saved from the applet help page goes here
 	 * @param force Force provider re initialization
 	 */
 	private EnsureProvider(force: boolean = false): void {
-		let currentName = this.provider?.name;
-		switch (this.config._dataService) {
-			case "DarkSky":           // No City Info
-				if (currentName != "DarkSky" || force) this.provider = new DarkSky(this);
-				break;
-			case "OpenWeatherMap":   // No City Info
-				if (currentName != "OpenWeatherMap" || force) this.provider = new OpenWeatherMap(this);
-				break;
-			case "MetNorway":         // No TZ or city info
-				if (currentName != "MetNorway" || force) this.provider = new MetNorway(this);
-				break;
-			case "Weatherbit":
-				if (currentName != "Weatherbit" || force) this.provider = new Weatherbit(this);
-				break;
-			case "Tomorrow.io":
-				if (currentName != "Tomorrow.io" || force) this.provider = new ClimacellV4(this);
-				break;
-			case "Met Office UK":
-				if (currentName != "Met Office UK" || force) this.provider = new MetUk(this);
-				break;
-			case "US Weather":
-				if (currentName != "US Weather" || force) this.provider = new USWeather(this);
-				break;
-			case "Visual Crossing":
-				if (currentName != "Visual Crossing" || force) this.provider = new VisualCrossing(this);
-				break;
-			case "DanishMI":
-				if (currentName != "DanishMI" || force) this.provider = new DanishMI(this);
-				break;
-			default:
-				Logger.Error(`Provider string "${currentName}" from settings doesn't exist, please contact the developer!`);
-				this.DisplayHardError("Critical Error", "Please see logs and contact developer");
-		}
+		const currentName = this.provider?.name;
+		if (currentName != this.config._dataService || force)
+			this.provider = ServiceClassMapping[this.config._dataService](this);
 	}
 
 	/** Fills in missing weather info from location Data
@@ -556,14 +525,14 @@ The contents of the file saved from the applet help page goes here
 		weatherInfo.condition.main = ProcessCondition(weatherInfo.condition.main, this.config._translateCondition);
 		weatherInfo.condition.description = ProcessCondition(weatherInfo.condition.description, this.config._translateCondition);
 
-		for (let index = 0; index < weatherInfo.forecasts.length; index++) {
-			const condition = weatherInfo.forecasts[index].condition;
+		for (const forecast of weatherInfo.forecasts) {
+			const condition = forecast.condition;
 			condition.main = ProcessCondition(condition.main, this.config._translateCondition);
 			condition.description = ProcessCondition(condition.description, this.config._translateCondition);
 		}
 
-		for (let index = 0; index < weatherInfo.hourlyForecasts.length; index++) {
-			const condition = weatherInfo.hourlyForecasts[index].condition;
+		for (const forecast of weatherInfo.hourlyForecasts) {
+			const condition = forecast.condition;
 			condition.main = ProcessCondition(condition.main, this.config._translateCondition);
 			condition.description = ProcessCondition(condition.description, this.config._translateCondition);
 		}
@@ -633,7 +602,7 @@ The contents of the file saved from the applet help page goes here
 			return;
 		}
 
-		let nextRefresh = this.loop.GetSecondsUntilNextRefresh();
+		const nextRefresh = this.loop.GetSecondsUntilNextRefresh();
 		Logger.Error("Retrying in the next " + nextRefresh.toString() + " seconds...");
 	}